// src/app/services/api/interfaces/prevenda/CS_IReqUpdateDD071.ts
export interface IReqUpdateDD071 {
    DD071_Id?: string
    DD070_ID?: string
    DD071_Tipo?: number
    DD071_Conta_ID?: string
    DD071_TipoDocto?: number
    DD071_CNPJ_CPF?: number
    DD071_IE_RG?: number
    DD071_Nome?: string
    DD071_Logradouro?: string
    DD071_Numero?: string
    DD071_Complemento?: string
    DD071_Perimetro?: string
    DD071_Bairro_ID?: string
    DD071_NomeBairro?: string
    DD071_CEP?: number
    DD071_Pais_ID?: string
    DD071_UF_ID?: string
    DD071_Cidade_ID?: string
    DD071_Telefone?: string
    DD071_Email?: string
    DD071_SUFRAMA?: string
    DD071_ModalidadeFrete?: number
    DD071_PlacaVeiculo?: string
    DD071_UFVeiculo?: string
    DD071_MarcaVeiculo?: string
    DD071_NumTransp?: string
    DD071_PlacaReboque?: string
    DD071_UFReboque_ID?: string
    DD071_NumTranspReboq?: string
    DD071_Vagao?: string
    DD071_Balsa?: string
    DD071_Endereco_ID?: string
    DD070_Send_Email?: boolean
    DD070_Send_SMS?: boolean
    DD070_User_Operador_ID?: string
    DD071_Data_Caixa?: string
    DD071_SMS?: string
    DD071_indFinal?: string
    dd071_Ident_Estrangeiro?: string
}